Make ASGIHandler and WSGIHandler extend genric TransportHandler
